Add relatorio de balanco e mudanças de tipagen

+++ ContabilidadeApi/Dto/LancamentoDebitoCreditoDto.cs
@@ -5,7 +5,7 @@ namespace ContabilidadeApi.Dto
     public class LancamentoDebitoCreditoDto
     {
         public DateTime Data { get; set; }
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
         public TipoOperacaoEnum TipoAcao { get; set; }  
         public string? DescComplementar { get; set; }
         public int ContaContabilId { get; set; }
